+++ main.tf
@@ -3,32 +3,15 @@ provider "azurerm" {
 }
 
 locals {
-  is_sharepoint_subscription = split("-", var.sharepoint_version)[0] == "Subscription" ? true : false
-  config_sp_image            = lookup(local.config_sp_image_list, split("-", var.sharepoint_version)[0])
-  config_sp_dsc              = local.is_sharepoint_subscription ? local.config_sp_se_dsc : local.config_sp_legacy_dsc
-  config_fe_dsc              = local.is_sharepoint_subscription ? local.config_fe_se_dsc : local.config_fe_legacy_dsc
-  create_rdp_rule            = lower(var.rdp_traffic_allowed) == "no" ? 0 : 1
+  resourceGroupNameFormatted = replace(replace(replace(replace(var.resource_group_name, ".", "-"), "(", "-"), ")", "-"), "_", "-")
   admin_password             = var.admin_password == "" ? random_password.random_admin_password.result : var.admin_password
   service_accounts_password  = var.service_accounts_password == "" ? random_password.random_service_accounts_password.result : var.service_accounts_password
+  create_rdp_rule            = lower(var.rdp_traffic_allowed) == "no" ? 0 : 1
   license_type               = var.enable_hybrid_benefit_server_licenses == true ? "Windows_Server" : "None"
-  sharepoint_bits_selected   = local.is_sharepoint_subscription ? jsonencode(local.sharepoint_subscription_bits) : jsonencode([])
   _artifactsLocation         = var._artifactsLocation
   _artifactsLocationSasToken = ""
-
-  general_settings = {
-    adfsSvcUserName       = "adfssvc"
-    sqlSvcUserName        = "sqlsvc"
-    spSetupUserName       = "spsetup"
-    spFarmUserName        = "spfarm"
-    spSvcUserName         = "spsvc"
-    spAppPoolUserName     = "spapppool"
-    spADDirSyncUserName   = "spdirsync"
-    spSuperUserName       = "spSuperUser"
-    spSuperReaderName     = "spSuperReader"
-    sqlAlias              = "SQLAlias"
-    bastion_publicip_name = "${lower(azurerm_resource_group.rg.name)}-bastion"
-  }
-
+  is_sharepoint_subscription = split("-", var.sharepoint_version)[0] == "Subscription" ? true : false
+  sharepoint_bits_used       = local.is_sharepoint_subscription ? jsonencode(local.sharepoint_subscription_bits) : jsonencode([])
   sharepoint_subscription_bits = [
     {
       "Label" : "RTM",
@@ -66,82 +49,57 @@ locals {
     vmDCPrivateIPAddress           = "10.1.1.4"
   }
 
-  config_dc = {
-    vmName             = "DC"
-    vmSize             = var.vm_dc_size
-    vmImagePublisher   = "MicrosoftWindowsServer"
-    vmImageOffer       = "WindowsServer"
-    vmImageSKU         = "2022-datacenter-azure-edition-smalldisk"
-    storageAccountType = var.vm_dc_storage_account_type
-  }
-
-  config_sql = {
-    vmName             = "SQL"
-    vmSize             = var.vm_sql_size
-    vmImagePublisher   = "MicrosoftSQLServer"
-    vmImageOffer       = "sql2019-ws2022"
-    vmImageSKU         = "sqldev-gen2"
-    storageAccountType = var.vm_sql_storage_account_type
-  }
-
-  config_sp = {
-    vmName             = "SP"
-    vmSize             = var.vm_sp_size
-    storageAccountType = var.vm_sp_storage_account_type
-  }
-
-  config_sp_image_list = {
-    "Subscription" = "MicrosoftWindowsServer:WindowsServer:2022-datacenter-azure-edition:latest"
-    "2019"         = "MicrosoftSharePoint:MicrosoftSharePointServer:sp2019:latest"
+  sharepoint_images_list = {
+    "Subscription" = "MicrosoftWindowsServer:WindowsServer:2022-datacenter-azure-edition-smalldisk:latest"
+    "2019"         = "MicrosoftSharePoint:MicrosoftSharePointServer:sp2019gen2smalldisk:latest"
     "2016"         = "MicrosoftSharePoint:MicrosoftSharePointServer:sp2016:latest"
     "2013"         = "MicrosoftSharePoint:MicrosoftSharePointServer:sp2013:latest"
   }
 
-  config_fe = {
-    vmName = "FE"
-    vmSize = var.vm_sp_size
+  vms_settings = {
+    vm_dc_name           = "DC"
+    vm_sql_name          = "SQL"
+    vm_sp_name           = "SP"
+    vm_fe_name           = "FE"
+    vm_dc_image          = "MicrosoftWindowsServer:WindowsServer:2022-datacenter-azure-edition-smalldisk:latest"
+    vm_sql_image         = "MicrosoftSQLServer:sql2022-ws2022:sqldev-gen2:latest"
+    vms_sharepoint_image = lookup(local.sharepoint_images_list, split("-", var.sharepoint_version)[0])
   }
 
-  config_dc_dsc = {
-    fileName       = "ConfigureDCVM.zip"
-    script         = "ConfigureDCVM.ps1"
-    function       = "ConfigureDCVM"
-    forceUpdateTag = "1.0"
+  dsc_settings = {
+    vm_dc_fileName = "ConfigureDCVM.zip"
+    vm_dc_script   = "ConfigureDCVM.ps1"
+    vm_dc_function = "ConfigureDCVM"
+
+    vm_sql_fileName = "ConfigureSQLVM.zip"
+    vm_sql_script   = "ConfigureSQLVM.ps1"
+    vm_sql_function = "ConfigureSQLVM"
+
+    vm_sp_fileName = local.is_sharepoint_subscription ? "ConfigureSPSE.zip" : "ConfigureSPLegacy.zip"
+    vm_sp_script   = local.is_sharepoint_subscription ? "ConfigureSPSE.ps1" : "ConfigureSPLegacy.ps1"
+    vm_sp_function = "ConfigureSPVM"
+
+    vm_fe_fileName = local.is_sharepoint_subscription ? "ConfigureFESE.zip" : "ConfigureFELegacy.zip"
+    vm_fe_script   = local.is_sharepoint_subscription ? "ConfigureFESE.ps1" : "ConfigureFELegacy.ps1"
+    vm_fe_function = "ConfigureFEVM"
   }
 
-  config_sql_dsc = {
-    fileName       = "ConfigureSQLVM.zip"
-    script         = "ConfigureSQLVM.ps1"
-    function       = "ConfigureSQLVM"
-    forceUpdateTag = "1.0"
-  }
-
-  config_sp_legacy_dsc = {
-    fileName       = "ConfigureSPLegacy.zip"
-    script         = "ConfigureSPLegacy.ps1"
-    function       = "ConfigureSPVM"
-    forceUpdateTag = "1.0"
-  }
-
-  config_sp_se_dsc = {
-    fileName       = "ConfigureSPSE.zip"
-    script         = "ConfigureSPSE.ps1"
-    function       = "ConfigureSPVM"
-    forceUpdateTag = "1.0"
-  }
-
-  config_fe_legacy_dsc = {
-    fileName       = "ConfigureFELegacy.zip"
-    script         = "ConfigureFELegacy.ps1"
-    function       = "ConfigureFEVM"
-    forceUpdateTag = "1.0"
-  }
-
-  config_fe_se_dsc = {
-    fileName       = "ConfigureFESE.zip"
-    script         = "ConfigureFESE.ps1"
-    function       = "ConfigureFEVM"
-    forceUpdateTag = "1.0"
+  deployment_settings = {
+    sharepoint_sites_authority    = "spsites"
+    sharepoint_central_admin_port = 5000
+    localAdminUserName            = "local-${var.admin_username}"
+    enable_analysis               = true # This enables a Python script that parses dsc logs on SharePoint VMs, to compute the time take by each resource to run
+    apply_browser_policies        = true
+    sqlAlias                      = "SQLAlias"
+    adfsSvcUserName               = "adfssvc"
+    sqlSvcUserName                = "sqlsvc"
+    spSetupUserName               = "spsetup"
+    spFarmUserName                = "spfarm"
+    spSvcUserName                 = "spsvc"
+    spAppPoolUserName             = "spapppool"
+    spADDirSyncUserName           = "spdirsync"
+    spSuperUserName               = "spSuperUser"
+    spSuperReaderName             = "spSuperReader"
   }
 }
 
@@ -173,7 +131,7 @@ resource "azurerm_resource_group" "rg" {
 
 # Create the network security groups
 resource "azurerm_network_security_group" "nsg_subnet_dc" {
-  name                = "NSG-Subnet-${local.config_dc["vmName"]}"
+  name                = "NSG-Subnet-${local.vms_settings.vm_dc_name}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 }
@@ -195,7 +153,7 @@ resource "azurerm_network_security_rule" "rdp_rule_subnet_dc" {
 }
 
 resource "azurerm_network_security_group" "nsg_subnet_sql" {
-  name                = "NSG-Subnet-${local.config_sql["vmName"]}"
+  name                = "NSG-Subnet-${local.vms_settings.vm_sql_name}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 }
@@ -217,7 +175,7 @@ resource "azurerm_network_security_rule" "rdp_rule_subnet_sql" {
 }
 
 resource "azurerm_network_security_group" "nsg_subnet_sp" {
-  name                = "NSG-Subnet-${local.config_sp["vmName"]}"
+  name                = "NSG-Subnet-${local.vms_settings.vm_sp_name}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 }
@@ -240,7 +198,7 @@ resource "azurerm_network_security_rule" "rdp_rule_subnet_sp" {
 
 # Create the virtual network, 3 subnets, and associate each subnet with its Network Security Group
 resource "azurerm_virtual_network" "vnet" {
-  name                = "${azurerm_resource_group.rg.name}-vnet"
+  name                = "${local.resourceGroupNameFormatted}-vnet"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
   address_space       = [local.network_settings["vNetPrivatePrefix"]]
@@ -248,7 +206,7 @@ resource "azurerm_virtual_network" "vnet" {
 
 # Subnet and NSG for DC
 resource "azurerm_subnet" "subnet_dc" {
-  name                 = "Subnet-${local.config_dc["vmName"]}"
+  name                 = "Subnet-${local.vms_settings.vm_dc_name}"
   resource_group_name  = azurerm_resource_group.rg.name
   virtual_network_name = azurerm_virtual_network.vnet.name
   address_prefixes     = [local.network_settings["vNetPrivateSubnetDCPrefix"]]
@@ -260,7 +218,7 @@ resource "azurerm_subnet_network_security_group_association" "nsg_subnetdc_assoc
 }
 
 resource "azurerm_subnet" "subnet_sql" {
-  name                 = "Subnet-${local.config_sql["vmName"]}"
+  name                 = "Subnet-${local.vms_settings.vm_sql_name}"
   resource_group_name  = azurerm_resource_group.rg.name
   virtual_network_name = azurerm_virtual_network.vnet.name
   address_prefixes     = [local.network_settings["vNetPrivateSubnetSQLPrefix"]]
@@ -272,7 +230,7 @@ resource "azurerm_subnet_network_security_group_association" "nsg_subnetsql_asso
 }
 
 resource "azurerm_subnet" "subnet_sp" {
-  name                 = "Subnet-${local.config_sp["vmName"]}"
+  name                 = "Subnet-${local.vms_settings.vm_sp_name}"
   resource_group_name  = azurerm_resource_group.rg.name
   virtual_network_name = azurerm_virtual_network.vnet.name
   address_prefixes     = [local.network_settings["vNetPrivateSubnetSPPrefix"]]
@@ -286,17 +244,17 @@ resource "azurerm_subnet_network_security_group_association" "nsg_subnetsp_assoc
 # Create artifacts for VM DC
 resource "azurerm_public_ip" "pip_dc" {
   count               = var.add_public_ip_address == "Yes" ? 1 : 0
-  name                = "PublicIP-${local.config_dc["vmName"]}"
+  name                = "PublicIP-${local.vms_settings.vm_dc_name}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
-  domain_name_label   = "${lower(var.resource_group_name)}-${lower(local.config_dc["vmName"])}"
+  domain_name_label   = "${lower(local.resourceGroupNameFormatted)}-${lower(local.vms_settings.vm_dc_name)}"
   allocation_method   = "Dynamic"
   sku                 = "Basic"
   sku_tier            = "Regional"
 }
 
 resource "azurerm_network_interface" "nic_dc_0" {
-  name                = "NIC-${local.config_dc["vmName"]}-0"
+  name                = "NIC-${local.vms_settings.vm_dc_name}-0"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 
@@ -312,17 +270,17 @@ resource "azurerm_network_interface" "nic_dc_0" {
 # Create artifacts for VM SQL
 resource "azurerm_public_ip" "pip_sql" {
   count               = var.add_public_ip_address == "Yes" ? 1 : 0
-  name                = "PublicIP-${local.config_sql["vmName"]}"
+  name                = "PublicIP-${local.vms_settings.vm_sql_name}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
-  domain_name_label   = "${lower(var.resource_group_name)}-${lower(local.config_sql["vmName"])}"
+  domain_name_label   = "${lower(local.resourceGroupNameFormatted)}-${lower(local.vms_settings.vm_sql_name)}"
   allocation_method   = "Dynamic"
   sku                 = "Basic"
   sku_tier            = "Regional"
 }
 
 resource "azurerm_network_interface" "nic_sql_0" {
-  name                = "NIC-${local.config_sql["vmName"]}-0"
+  name                = "NIC-${local.vms_settings.vm_sql_name}-0"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 
@@ -337,17 +295,17 @@ resource "azurerm_network_interface" "nic_sql_0" {
 # Create artifacts for VM SP
 resource "azurerm_public_ip" "pip_sp" {
   count               = var.add_public_ip_address == "Yes" || var.add_public_ip_address == "SharePointVMsOnly" ? 1 : 0
-  name                = "PublicIP-${local.config_sp["vmName"]}"
+  name                = "PublicIP-${local.vms_settings.vm_sp_name}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
-  domain_name_label   = "${lower(var.resource_group_name)}-${lower(local.config_sp["vmName"])}"
+  domain_name_label   = "${lower(local.resourceGroupNameFormatted)}-${lower(local.vms_settings.vm_sp_name)}"
   allocation_method   = "Dynamic"
   sku                 = "Basic"
   sku_tier            = "Regional"
 }
 
 resource "azurerm_network_interface" "nic_sp_0" {
-  name                = "NIC-${local.config_sp["vmName"]}-0"
+  name                = "NIC-${local.vms_settings.vm_sp_name}-0"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 
@@ -361,12 +319,12 @@ resource "azurerm_network_interface" "nic_sp_0" {
 
 # Create virtual machines
 resource "azurerm_windows_virtual_machine" "vm_dc" {
-  name                     = local.config_dc["vmName"]
-  computer_name            = local.config_dc["vmName"]
+  name                     = local.vms_settings.vm_dc_name
+  computer_name            = local.vms_settings.vm_dc_name
   location                 = azurerm_resource_group.rg.location
   resource_group_name      = azurerm_resource_group.rg.name
   network_interface_ids    = [azurerm_network_interface.nic_dc_0.id]
-  size                     = local.config_dc["vmSize"]
+  size                     = var.vm_dc_size
   admin_username           = var.admin_username
   admin_password           = local.admin_password
   license_type             = local.license_type
@@ -375,22 +333,22 @@ resource "azurerm_windows_virtual_machine" "vm_dc" {
   provision_vm_agent       = true
 
   os_disk {
-    name                 = "Disk-${local.config_dc["vmName"]}-OS"
-    storage_account_type = local.config_dc["storageAccountType"]
+    name                 = "Disk-${local.vms_settings.vm_dc_name}-OS"
+    storage_account_type = var.vm_dc_storage_account_type
     caching              = "ReadWrite"
   }
 
   source_image_reference {
-    publisher = local.config_dc["vmImagePublisher"]
-    offer     = local.config_dc["vmImageOffer"]
-    sku       = local.config_dc["vmImageSKU"]
-    version   = "latest"
+    publisher = split(":", local.vms_settings.vm_dc_image)[0]
+    offer     = split(":", local.vms_settings.vm_dc_image)[1]
+    sku       = split(":", local.vms_settings.vm_dc_image)[2]
+    version   = split(":", local.vms_settings.vm_dc_image)[3]
   }
 }
 
 resource "azurerm_virtual_machine_extension" "vm_dc_dsc" {
   # count                      = 0
-  name                       = "VM-${local.config_dc["vmName"]}-DSC"
+  name                       = "VM-${local.vms_settings.vm_dc_name}-DSC"
   virtual_machine_id         = azurerm_windows_virtual_machine.vm_dc.id
   publisher                  = "Microsoft.Powershell"
   type                       = "DSC"
@@ -405,13 +363,17 @@ resource "azurerm_virtual_machine_extension" "vm_dc_dsc" {
   {
     "wmfVersion": "latest",
     "configuration": {
-	    "url": "${local._artifactsLocation}${local.config_dc_dsc["fileName"]}${local._artifactsLocationSasToken}",
-	    "function": "${local.config_dc_dsc["function"]}",
-	    "script": "${local.config_dc_dsc["script"]}"
+	    "url": "${local._artifactsLocation}${local.dsc_settings["vm_dc_fileName"]}${local._artifactsLocationSasToken}",
+	    "function": "${local.dsc_settings["vm_dc_function"]}",
+	    "script": "${local.dsc_settings["vm_dc_script"]}"
     },
     "configurationArguments": {
       "domainFQDN": "${var.domain_fqdn}",
-      "PrivateIP": "${local.network_settings["vmDCPrivateIPAddress"]}"
+      "PrivateIP": "${local.network_settings["vmDCPrivateIPAddress"]}",
+      "SPServerName": "${local.vms_settings.vm_sp_name}",
+      "SharePointSitesAuthority": "${local.deployment_settings.sharepoint_sites_authority}",
+      "SharePointCentralAdminPort": "${local.deployment_settings.sharepoint_central_admin_port}",
+      "ApplyBrowserPolicies": ${local.deployment_settings.apply_browser_policies}
     },
     "privacy": {
       "dataCollection": "enable"
@@ -427,7 +389,7 @@ SETTINGS
         "Password": "${local.admin_password}"
       },
       "AdfsSvcCreds": {
-        "UserName": "${local.general_settings["adfsSvcUserName"]}",
+        "UserName": "${local.deployment_settings.adfsSvcUserName}",
         "Password": "${local.service_accounts_password}"
       }
     }
@@ -450,13 +412,13 @@ resource "azurerm_dev_test_global_vm_shutdown_schedule" "vm_dc_shutdown" {
 }
 
 resource "azurerm_windows_virtual_machine" "vm_sql" {
-  name                     = local.config_sql["vmName"]
-  computer_name            = local.config_sql["vmName"]
+  name                     = local.vms_settings.vm_sql_name
+  computer_name            = local.vms_settings.vm_sql_name
   location                 = azurerm_resource_group.rg.location
   resource_group_name      = azurerm_resource_group.rg.name
   network_interface_ids    = [azurerm_network_interface.nic_sql_0.id]
-  size                     = local.config_sql["vmSize"]
-  admin_username           = "local-${var.admin_username}"
+  size                     = var.vm_sql_size
+  admin_username           = local.deployment_settings.localAdminUserName
   admin_password           = local.admin_password
   license_type             = local.license_type
   timezone                 = var.time_zone
@@ -464,22 +426,22 @@ resource "azurerm_windows_virtual_machine" "vm_sql" {
   provision_vm_agent       = true
 
   os_disk {
-    name                 = "Disk-${local.config_sql["vmName"]}-OS"
-    storage_account_type = local.config_sql["storageAccountType"]
+    name                 = "Disk-${local.vms_settings.vm_sql_name}-OS"
+    storage_account_type = var.vm_sql_storage_account_type
     caching              = "ReadWrite"
   }
 
   source_image_reference {
-    publisher = local.config_sql["vmImagePublisher"]
-    offer     = local.config_sql["vmImageOffer"]
-    sku       = local.config_sql["vmImageSKU"]
-    version   = "latest"
+    publisher = split(":", local.vms_settings.vm_sql_image)[0]
+    offer     = split(":", local.vms_settings.vm_sql_image)[1]
+    sku       = split(":", local.vms_settings.vm_sql_image)[2]
+    version   = split(":", local.vms_settings.vm_sql_image)[3]
   }
 }
 
 resource "azurerm_virtual_machine_extension" "vm_sql_dsc" {
   # count                      = 0
-  name                       = "VM-${local.config_sql["vmName"]}-DSC"
+  name                       = "VM-${local.vms_settings.vm_sql_name}-DSC"
   virtual_machine_id         = azurerm_windows_virtual_machine.vm_sql.id
   publisher                  = "Microsoft.Powershell"
   type                       = "DSC"
@@ -494,12 +456,12 @@ resource "azurerm_virtual_machine_extension" "vm_sql_dsc" {
   {
     "wmfVersion": "latest",
     "configuration": {
-	    "url": "${local._artifactsLocation}${local.config_sql_dsc["fileName"]}${local._artifactsLocationSasToken}",
-	    "function": "${local.config_sql_dsc["function"]}",
-	    "script": "${local.config_sql_dsc["script"]}"
+	    "url": "${local._artifactsLocation}${local.dsc_settings["vm_sql_fileName"]}${local._artifactsLocationSasToken}",
+	    "function": "${local.dsc_settings["vm_sql_function"]}",
+	    "script": "${local.dsc_settings["vm_sql_script"]}"
     },
     "configurationArguments": {
-      "DNSServer": "${local.network_settings["vmDCPrivateIPAddress"]}",
+      "DNSServerIP": "${local.network_settings["vmDCPrivateIPAddress"]}",
       "DomainFQDN": "${var.domain_fqdn}"
     },
     "privacy": {
@@ -516,11 +478,11 @@ SETTINGS
         "Password": "${local.admin_password}"
       },
       "SqlSvcCreds": {
-        "UserName": "${local.general_settings["sqlSvcUserName"]}",
+        "UserName": "${local.deployment_settings.sqlSvcUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPSetupCreds": {
-        "UserName": "${local.general_settings["spSetupUserName"]}",
+        "UserName": "${local.deployment_settings.spSetupUserName}",
         "Password": "${local.service_accounts_password}"
       }
     }
@@ -543,13 +505,13 @@ resource "azurerm_dev_test_global_vm_shutdown_schedule" "vm_sql_shutdown" {
 }
 
 resource "azurerm_windows_virtual_machine" "vm_sp" {
-  name                     = local.config_sp["vmName"]
-  computer_name            = local.config_sp["vmName"]
+  name                     = local.vms_settings.vm_sp_name
+  computer_name            = local.vms_settings.vm_sp_name
   location                 = azurerm_resource_group.rg.location
   resource_group_name      = azurerm_resource_group.rg.name
   network_interface_ids    = [azurerm_network_interface.nic_sp_0.id]
-  size                     = local.config_sp["vmSize"]
-  admin_username           = "local-${var.admin_username}"
+  size                     = var.vm_sp_size
+  admin_username           = local.deployment_settings.localAdminUserName
   admin_password           = local.admin_password
   license_type             = local.license_type
   timezone                 = var.time_zone
@@ -557,22 +519,22 @@ resource "azurerm_windows_virtual_machine" "vm_sp" {
   provision_vm_agent       = true
 
   os_disk {
-    name                 = "Disk-${local.config_sp["vmName"]}-OS"
-    storage_account_type = local.config_sp["storageAccountType"]
+    name                 = "Disk-${local.vms_settings.vm_sp_name}-OS"
+    storage_account_type = var.vm_sp_storage_account_type
     caching              = "ReadWrite"
   }
 
   source_image_reference {
-    publisher = split(":", local.config_sp_image)[0]
-    offer     = split(":", local.config_sp_image)[1]
-    sku       = split(":", local.config_sp_image)[2]
-    version   = "latest"
+    publisher = split(":", local.vms_settings.vms_sharepoint_image)[0]
+    offer     = split(":", local.vms_settings.vms_sharepoint_image)[1]
+    sku       = split(":", local.vms_settings.vms_sharepoint_image)[2]
+    version   = split(":", local.vms_settings.vms_sharepoint_image)[3]
   }
 }
 
 resource "azurerm_virtual_machine_extension" "vm_sp_dsc" {
   # count                      = 0
-  name                       = "VM-${local.config_sp["vmName"]}-DSC"
+  name                       = "VM-${local.vms_settings.vm_sp_name}-DSC"
   virtual_machine_id         = azurerm_windows_virtual_machine.vm_sp.id
   publisher                  = "Microsoft.Powershell"
   type                       = "DSC"
@@ -587,19 +549,21 @@ resource "azurerm_virtual_machine_extension" "vm_sp_dsc" {
   {
     "wmfVersion": "latest",
     "configuration": {
-	    "url": "${local._artifactsLocation}${local.config_sp_dsc["fileName"]}${local._artifactsLocationSasToken}",
-	    "function": "${local.config_sp_dsc["function"]}",
-	    "script": "${local.config_sp_dsc["script"]}"
+	    "url": "${local._artifactsLocation}${local.dsc_settings["vm_sp_fileName"]}${local._artifactsLocationSasToken}",
+	    "function": "${local.dsc_settings["vm_sp_function"]}",
+	    "script": "${local.dsc_settings["vm_sp_script"]}"
     },
     "configurationArguments": {
-      "DNSServer": "${local.network_settings["vmDCPrivateIPAddress"]}",
+      "DNSServerIP": "${local.network_settings["vmDCPrivateIPAddress"]}",
       "DomainFQDN": "${var.domain_fqdn}",
-      "DCName": "${local.config_dc["vmName"]}",
-      "SQLName": "${local.config_sql["vmName"]}",
-      "SQLAlias": "${local.general_settings["sqlAlias"]}",
+      "DCServerName": "${local.vms_settings.vm_dc_name}",
+      "SQLServerName": "${local.vms_settings.vm_sql_name}",
+      "SQLAlias": "${local.deployment_settings.sqlAlias}",
       "SharePointVersion": "${var.sharepoint_version}",
-      "EnableAnalysis": false,
-      "SharePointBits": ${local.sharepoint_bits_selected}
+      "SharePointSitesAuthority": "${local.deployment_settings.sharepoint_sites_authority}",
+      "SharePointCentralAdminPort": "${local.deployment_settings.sharepoint_central_admin_port}",
+      "EnableAnalysis": ${local.deployment_settings.enable_analysis},
+      "SharePointBits": ${local.sharepoint_bits_used}
     },
     "privacy": {
       "dataCollection": "enable"
@@ -615,23 +579,23 @@ SETTINGS
         "Password": "${local.admin_password}"
       },
       "SPSetupCreds": {
-        "UserName": "${local.general_settings["spSetupUserName"]}",
+        "UserName": "${local.deployment_settings.spSetupUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPFarmCreds": {
-        "UserName": "${local.general_settings["spFarmUserName"]}",
+        "UserName": "${local.deployment_settings.spFarmUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPSvcCreds": {
-        "UserName": "${local.general_settings["spSvcUserName"]}",
+        "UserName": "${local.deployment_settings.spSvcUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPAppPoolCreds": {
-        "UserName": "${local.general_settings["spAppPoolUserName"]}",
+        "UserName": "${local.deployment_settings.spAppPoolUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPADDirSyncCreds": {
-        "UserName": "${local.general_settings["spADDirSyncUserName"]}",
+        "UserName": "${local.deployment_settings.spADDirSyncUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPPassphraseCreds": {
@@ -639,11 +603,11 @@ SETTINGS
         "Password": "${local.service_accounts_password}"
       },
       "SPSuperUserCreds": {
-        "UserName": "${local.general_settings["spSuperUserName"]}",
+        "UserName": "${local.deployment_settings.spSuperUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPSuperReaderCreds": {
-        "UserName": "${local.general_settings["spSuperReaderName"]}",
+        "UserName": "${local.deployment_settings.spSuperReaderName}",
         "Password": "${local.service_accounts_password}"
       }
     }
@@ -668,10 +632,10 @@ resource "azurerm_dev_test_global_vm_shutdown_schedule" "vm_sp_shutdown" {
 # Can create 0 to var.number_additional_frontend FE VMs
 resource "azurerm_public_ip" "pip_fe" {
   count               = var.add_public_ip_address == "Yes" || var.add_public_ip_address == "SharePointVMsOnly" ? var.number_additional_frontend : 0
-  name                = "PublicIP-${local.config_fe["vmName"]}-${count.index}"
+  name                = "PublicIP-${local.vms_settings.vm_fe_name}-${count.index}"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
-  domain_name_label   = "${lower(var.resource_group_name)}-${lower(local.config_fe["vmName"])}-${count.index}"
+  domain_name_label   = "${lower(local.resourceGroupNameFormatted)}-${lower(local.vms_settings.vm_fe_name)}-${count.index}"
   allocation_method   = "Dynamic"
   sku                 = "Basic"
   sku_tier            = "Regional"
@@ -679,7 +643,7 @@ resource "azurerm_public_ip" "pip_fe" {
 
 resource "azurerm_network_interface" "nic_fe_0" {
   count               = var.number_additional_frontend
-  name                = "NIC-${local.config_fe["vmName"]}-${count.index}-0"
+  name                = "NIC-${local.vms_settings.vm_fe_name}-${count.index}-0"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
 
@@ -693,13 +657,13 @@ resource "azurerm_network_interface" "nic_fe_0" {
 
 resource "azurerm_windows_virtual_machine" "vm_fe" {
   count                    = var.number_additional_frontend
-  name                     = "${local.config_fe["vmName"]}-${count.index}"
-  computer_name            = "${local.config_fe["vmName"]}-${count.index}"
+  name                     = "${local.vms_settings.vm_fe_name}-${count.index}"
+  computer_name            = "${local.vms_settings.vm_fe_name}-${count.index}"
   location                 = azurerm_resource_group.rg.location
   resource_group_name      = azurerm_resource_group.rg.name
   network_interface_ids    = [element(azurerm_network_interface.nic_fe_0.*.id, count.index)]
-  size                     = local.config_sp["vmSize"]
-  admin_username           = "local-${var.admin_username}"
+  size                     = var.vm_sp_size
+  admin_username           = local.deployment_settings.localAdminUserName
   admin_password           = local.admin_password
   license_type             = local.license_type
   timezone                 = var.time_zone
@@ -707,23 +671,23 @@ resource "azurerm_windows_virtual_machine" "vm_fe" {
   provision_vm_agent       = true
 
   os_disk {
-    name                 = "Disk-${local.config_fe["vmName"]}-${count.index}-OS"
-    storage_account_type = local.config_sp["storageAccountType"]
+    name                 = "Disk-${local.vms_settings.vm_fe_name}-${count.index}-OS"
+    storage_account_type = var.vm_sp_storage_account_type
     caching              = "ReadWrite"
   }
 
   source_image_reference {
-    publisher = split(":", local.config_sp_image)[0]
-    offer     = split(":", local.config_sp_image)[1]
-    sku       = split(":", local.config_sp_image)[2]
-    version   = "latest"
+    publisher = split(":", local.vms_settings.vms_sharepoint_image)[0]
+    offer     = split(":", local.vms_settings.vms_sharepoint_image)[1]
+    sku       = split(":", local.vms_settings.vms_sharepoint_image)[2]
+    version   = split(":", local.vms_settings.vms_sharepoint_image)[3]
   }
 }
 
 resource "azurerm_virtual_machine_extension" "vm_fe_dsc" {
   # count                      = 0
   count                      = var.number_additional_frontend
-  name                       = "VM-${local.config_fe["vmName"]}-${count.index}-DSC"
+  name                       = "VM-${local.vms_settings.vm_fe_name}-${count.index}-DSC"
   virtual_machine_id         = element(azurerm_windows_virtual_machine.vm_fe.*.id, count.index)
   publisher                  = "Microsoft.Powershell"
   type                       = "DSC"
@@ -738,19 +702,20 @@ resource "azurerm_virtual_machine_extension" "vm_fe_dsc" {
   {
     "wmfVersion": "latest",
     "configuration": {
-	    "url": "${local._artifactsLocation}${local.config_fe_dsc["fileName"]}${local._artifactsLocationSasToken}",
-	    "function": "${local.config_fe_dsc["function"]}",
-	    "script": "${local.config_fe_dsc["script"]}"
+	    "url": "${local._artifactsLocation}${local.dsc_settings["vm_fe_fileName"]}${local._artifactsLocationSasToken}",
+	    "function": "${local.dsc_settings["vm_fe_function"]}",
+	    "script": "${local.dsc_settings["vm_fe_script"]}"
     },
     "configurationArguments": {
-      "DNSServer": "${local.network_settings["vmDCPrivateIPAddress"]}",
+      "DNSServerIP": "${local.network_settings["vmDCPrivateIPAddress"]}",
       "DomainFQDN": "${var.domain_fqdn}",
-      "DCName": "${local.config_dc["vmName"]}",
-      "SQLName": "${local.config_sql["vmName"]}",
-      "SQLAlias": "${local.general_settings["sqlAlias"]}",
+      "DCServerName": "${local.vms_settings.vm_dc_name}",
+      "SQLServerName": "${local.vms_settings.vm_sql_name}",
+      "SQLAlias": "${local.deployment_settings.sqlAlias}",
       "SharePointVersion": "${var.sharepoint_version}",
-      "EnableAnalysis": false,
-      "SharePointBits": ${local.sharepoint_bits_selected}
+      "SharePointSitesAuthority": "${local.deployment_settings.sharepoint_sites_authority}",
+      "EnableAnalysis": ${local.deployment_settings.enable_analysis},
+      "SharePointBits": ${local.sharepoint_bits_used}
     },
     "privacy": {
       "dataCollection": "enable"
@@ -766,11 +731,11 @@ SETTINGS
         "Password": "${local.admin_password}"
       },
       "SPSetupCreds": {
-        "UserName": "${local.general_settings["spSetupUserName"]}",
+        "UserName": "${local.deployment_settings.spSetupUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPFarmCreds": {
-        "UserName": "${local.general_settings["spFarmUserName"]}",
+        "UserName": "${local.deployment_settings.spFarmUserName}",
         "Password": "${local.service_accounts_password}"
       },
       "SPPassphraseCreds": {
@@ -983,7 +948,7 @@ resource "azurerm_public_ip" "pip_bastion" {
   name                = "PublicIP-Bastion"
   location            = azurerm_resource_group.rg.location
   resource_group_name = azurerm_resource_group.rg.name
-  domain_name_label   = local.general_settings.bastion_publicip_name
+  domain_name_label   = "${lower(local.resourceGroupNameFormatted)}-bastion"
   allocation_method   = "Static"
   sku                 = "Standard"
   sku_tier            = "Regional"